[RFC] GraphiQL when loaded in browser
This RFC adds an opt-in behavior of express-graphql to render a fully functional GraphiQL instance when loaded from a browser - determed as anything that prefers HTML over JSON from the Accept header.

GraphiQL mode is enabled by providing `graphiql: true` in the options bag. It defaults to `false` which is a reasonable default, but also ensures any existing use updating to this version will not unknowingly add this feature. The examples in the README have been changed to enable GraphiQL, ideally all future users will leave this on, cargo-cult from README is likely.

When GraphiQL is shown, the query result is run on the server first and presented immediately in the GraphiQL UI.

If you wish to use the browser as a cheap cURL, just add `&raw` to the end of the URL and GraphiQL is replaced by raw JSON output.

One small detail: any non-GraphQL GET params provided will be carried through by GraphiQL when executing queries. Ideally this covers the majority of authentication cases where are likely to look something like `https://mysite/graphql?oauth=abc123&query={yo}`.