Pass request to extensionFn
in my project, i remove fields in the extension function based on the requesting user. since it is common to store the user on the request object, i thought this might be useful for others. i do not remove the fields in the resolve function so that child-resolves have access to a fully-hydrated parent resource.