Fixes and improvements to the README (#748)
* Fix rootValue option description

The mentioned function should be execute() instead of graphql(). Also,
the branch name and line number have changed.

* Fix context option description

The mentioned function should be execute() instead of graphql() and the
passed property should be contextValue instead of context. Also, the
branch name and line number have changed.

* Remove distracting session middleware from extensions example

Using the session middleware in the extensions example seems unnecessary
and distracting from the extensions example.

* Make list items explaining POST body parsing consistent

* Add missing comma in JSON example

* Improve the readability of some sentences

* Fix some small typos

* Add missing backticks

* Fix branch name in links to graphql-js repo